requirements.txt
google-auth>=2.0
google-api-python-client>=2.0
pandas>=2.0
openpyxl>=3.0
Pillow>=10.0
